microgo/pyproject.toml
[tool.poetry]
name = "microgo"
version = "0.1.0"
description = ""
authors = ["highlander-spirou <highlandetspirou2812@gmail.com>"]
readme = "README.md"
packages = [
    { include = "auth", from="lib" },
    { include = "errors", from="lib" },
    { include = "file_access", from="lib" },
    { include = "log", from="lib" }
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.24"
mysql-connector-python = "^8.2.0"
pymongo = "^4.6.1"
pytz = "^2023.3.post1"
gunicorn = "^21.2.0"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
